[00447] Extract the Review folder gate into PlanRefreshGate
The check that drops a PlansChanged event naming another plan lived as a private helper on the Review content view, so the other single-plan view could not reuse it.
